refactor: break the "halo2_frontend/src/circuit.rs" file
